Support multiple fleets per airline
Each airline may specify different fleets by including a 'fleets' object
with each key being a weighted list of aircraft in the same format as the
existing 'aircraft' key.  A 'default' fleet should be specified, if it is
not specified the 'aircraft' list will be used.

Airports may specify a specific fleet to use from a given airline by
appending a '/name' to the airline name in arrivals and departures.  If
no fleet is specified the default fleet is used.  For example to select
long haul aircraft from American the designator 'AAL/long' can be used.