New Airport Help - KCVG
To start off: I go to a boarding school with no access to outside PCs
until May. The laptop I do have access to is restricted from installing
anything. So, I can't install NodeJS. I would really like to contribute
to the game, though. So, I've created a file for the KCVG airport (my
city's airport) and filled out as much of the information as I could,
except the SIDs and the Spawn Patterns.

I used Brussels as a template and just replaced everything with CVG
data. I believe I may have done the fixes incorrectly, along with the
SIDs and Spawn Patterns not being completed, but I can't test the
airport OR figure out how to improve it any more than this.

If you guys have any critiques or recommendations for me to improve
this, please, let me know.

I would also appreciate it if any of you could test it for me,
considering that I can't test it for myself, and let me know what to
fix.

I'm planning on learning from this, so that when I add more airports in
the future, I can be more reliant on myself instead of you guys.

Thanks!