refactor/727 - update message substitutions
The tutorial messages have a number of placeholder strings for
replacement with actual data (say, the callsign of the a/c under
control).

This commit adds a few for where they're needed (taxi/ils), and removes
a bunch where they had been copy-pasted when making tutorial steps, but
they were not in use.